UploadForm: add (and wire up) a dataset type config prop
