--- src/lib/Utils/upload-config.ts
@@ -1,26 +1,0 @@
-import { intersection } from 'lodash';
-
-import { DatasetUploadPageConfig, DatasetUploadTypeConfig } from './types';
-
-export const uploadTypeConfig = {} as const;
-
-export function makeDatasetUploadPageConfig<
-  T1 extends string,
-  T2 extends string
->(
-  availableUploadTypes: T1[] = [],
-  uploadTypeConfig: DatasetUploadTypeConfig<T2>
-): DatasetUploadPageConfig<T1 & T2, T2> {
-  const restrictedUploadTypes = intersection(
-    availableUploadTypes,
-    Object.keys(uploadTypeConfig)
-  ) as (T1 & T2)[];
-
-  return restrictedUploadTypes.length === 0
-    ? { hasDirectUpload: false }
-    : {
-        hasDirectUpload: true,
-        availableUploadTypes: restrictedUploadTypes,
-        uploadTypeConfig,
-      };
-}
